# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Church Management APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/sacraments": {
            "get": {
                "tags": [
                    "Sacrament"
                ],
                "summary": "getSacramentList",
                "description": "Get all Sacraments",
                "operationId": "fe58f1e5945882bdf02d7241a640ccfa",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Sacrament"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sacrament"
                ],
                "summary": "createSacrament",
                "description": "Create Sacrament",
                "operationId": "5421d51fe9a3cae8e1d997a95627aabc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Sacrament"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Sacrament"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/sacraments/{id}": {
            "get": {
                "tags": [
                    "Sacrament"
                ],
                "summary": "getSacramentItem",
                "description": "Get Sacrament",
                "operationId": "747ad6770a96f2b203df4464e2c44e92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Sacrament",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Sacrament"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sacrament"
                ],
                "summary": "updateSacrament",
                "description": "Update Sacrament",
                "operationId": "fc5dd34569036e796fd0c48e06c23381",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Sacrament",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Sacrament"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Sacrament"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sacrament"
                ],
                "summary": "deleteSacrament",
                "description": "Delete Sacrament",
                "operationId": "8c98eeaa6233d4a57ff27bd289322d10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Sacrament",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Sacrament": {
                "required": [
                    "reason",
                    "amount",
                    "sacramented_at"
                ],
                "properties": {
                    "reason": {
                        "description": "Motif du sacrement|in:Confirmation,Baptême,Première communion,Mariage,Ordre,Onction des malades'",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "amount": {
                        "description": "",
                        "type": "integer",
                        "format": "int32",
                        "readOnly": false,
                        "nullable": false
                    },
                    "user_id": {
                        "description": "",
                        "type": "integer",
                        "format": "int32",
                        "readOnly": false,
                        "nullable": true
                    },
                    "sacramented_at": {
                        "description": "Date du sacrement",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": false,
                        "nullable": false
                    },
                    "created_at": {
                        "description": "Date d'enregistrement",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "updated_at": {
                        "description": "",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}